feat(tui): add exponential moving average for network speed display
- Implement smoothed network speed calculation using EMA with alpha=0.3
- Move TrackingConn from separate file to conn.go for better organization
- Update format specifier from %.2f to %.1f for cleaner display
- Improve error messages in route_unsupported.go with consistent quoting
